arm64: Add NEEDS_CLANG_HACKS to gate Google's 5.0 hacks
This is fixed with the Clang 6.0 builds available on Google's master
branch but it's just easier to have users compile with the 5.0 builds
that come with the 8.1 tags than to have them track that. I'll set
this flag in my scripts.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
(cherry picked from commit 7debf9573578e846df3de802054bbd6b61a6e2d6)